+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 ==========
 
+v1.5 (18/7/14)
+--------------
+* code updated to support the new Google Oauth2 API as the one used by the plugin is been deprecated in two months.  
+See: https://developers.google.com/+/api/auth-migration#timetable
+
+*ACTION REQUIRED* you need to enable the Google+ API in the Google developer console:  
+Project > API & AUTH > APIs
+
 v1.5
 ----
 * add user_loggedin event - it uses Events 2 introduced in Moodle 2.6